Don't add soft constraints to the active constraints. Fix some smaller stuff.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@40339 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/libs/linprog/ActiveSetSolver.cpp
@@ -369,7 +369,7 @@ ActiveSetSolver::Solve()
 		system.B(rowIndex) = constraint->RightSide();
 		for (int32 sIndex = 0; sIndex < leftSide->CountItems(); sIndex++ ) {
 			Summand* summand = leftSide->ItemAt(sIndex);
-			int32 coefficient = summand->Coeff();
+			double coefficient = summand->Coeff();
 			system.A(rowIndex, summand->VariableIndex()) = coefficient;
 		}
 		if (constraint->Op() == kLE) {
